refactor: commons 디렉토리와 common 디렉토리 통합
