+++ src/components/home/Restaurant.tsx
@@ -1,5 +1,5 @@
 import { css } from '@emotion/react';
-import { Col } from '../commons/Flex';
+import { Col } from '../common/Flex';
 import { useState } from 'react';
 import StoreRanking from './StoreRanking';
 import TitleSection from './TitleSection';
