+++ src/components/home/StoreRanking.tsx
@@ -1,5 +1,5 @@
 import { css } from '@emotion/react';
-import { Col, Row } from '../commons/Flex';
+import { Col, Row } from '../common/Flex';
 import starFilledImg from '../../assets/home/starFilled.svg';
 import starOutlinedImg from '../../assets/home/starOutlined.svg';
