+++ src/components/initialSettings/InitialSettingsCardContainer.tsx
@@ -1,6 +1,6 @@
 /** @jsxImportSource @emotion/react */
 import { css } from '@emotion/react';
-import { Col } from '../commons/Flex';
+import { Col } from '../common/Flex';
 
 import { ReactNode } from 'react';
