+++ src/components/makeGroup/NotificationStep.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
-import { Col } from '../commons/Flex';
+import { Col } from '../common/Flex';
 import Button from '../common/Button';
-import BackHeader from '../commons/BackHeader';
+import BackHeader from '../common/BackHeader';
 import { css } from '@emotion/react';
 import { messages } from '../../constants/dummyData';
