+++ src/components/restaurantList/MuckpotJoinModal.tsx
@@ -1,5 +1,5 @@
 import { css } from '@emotion/react';
-import { Col } from '../commons/Flex';
+import { Col } from '../common/Flex';
 import MuckpotInfoBox from './MuckpotInfoBox';
 
 interface Muckpot {
