+++ src/pages/Login.tsx
@@ -1,6 +1,6 @@
 /** @jsxImportSource @emotion/react */
 import { css } from '@emotion/react';
-import { Col } from '../components/commons/Flex';
+import { Col } from '../components/common/Flex';
 import { useNavigate } from 'react-router-dom';
 import { useState } from 'react';
