+++ src/pages/register/RegisterAccount.tsx
@@ -1,8 +1,8 @@
 /** @jsxImportSource @emotion/react */
 import { css } from '@emotion/react';
 import { useState, useEffect } from 'react';
-import { Col } from '../../components/commons/Flex';
-import BackHeader from '../../components/commons/BackHeader';
+import { Col } from '../../components/common/Flex';
+import BackHeader from '../../components/common/BackHeader';
 import Button from '../../components/common/Button';
 import { useNavigate } from 'react-router-dom';
