+++ tsconfig.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/app.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/components/common/backbutton.tsx","./src/components/common/bigbutton.tsx","./src/components/common/button.tsx","./src/components/common/dropdown.tsx","./src/components/common/searchbar.tsx","./src/components/common/searchinput.tsx","./src/components/commons/agreementbutton.tsx","./src/components/commons/backheader.tsx","./src/components/commons/bottomnavigation.tsx","./src/components/commons/checkbox.tsx","./src/components/commons/flex.tsx","./src/components/commons/header.tsx","./src/components/commons/searchdropdown.tsx","./src/components/home/cafe.tsx","./src/components/home/homecard.tsx","./src/components/home/restaurant.tsx","./src/components/home/storeranking.tsx","./src/components/home/taglist.tsx","./src/components/home/titlesection.tsx","./src/components/home/visitedrestaurant.tsx","./src/components/initialsettings/initialsettingscardcontainer.tsx","./src/components/initialsettings/initialsettingsheader.tsx","./src/components/initialsettings/setcafebrandcard.tsx","./src/components/initialsettings/setdepartmentcard.tsx","./src/components/initialsettings/setdrinkfeaturecard.tsx","./src/components/initialsettings/setgendercard.tsx","./src/components/initialsettings/setnamecard.tsx","./src/components/initialsettings/setrestaurantfeaturecard.tsx","./src/components/initialsettings/setrestauranttypecard.tsx","./src/components/makegroup/completestep.tsx","./src/components/makegroup/confirmstep.tsx","./src/components/makegroup/notificationstep.tsx","./src/components/makegroup/participantsstep.tsx","./src/components/makegroup/schedulestep.tsx","./src/components/makegroup/titlestep.tsx","./src/components/mypage/contentbox.tsx","./src/components/restaurantlist/makemuckpotbutton.tsx","./src/components/restaurantlist/muckpotcreatemodal.tsx","./src/components/restaurantlist/muckpotinfobox.tsx","./src/components/restaurantlist/muckpotjoinmodal.tsx","./src/components/restaurantlist/tagfilter.tsx","./src/constants/brands.ts","./src/constants/dummydata.ts","./src/firebase/firebase.ts","./src/pages/cafelist.tsx","./src/pages/findemailandpassword.tsx","./src/pages/group.tsx","./src/pages/home.tsx","./src/pages/initialsettings.tsx","./src/pages/login.tsx","./src/pages/makegroup.tsx","./src/pages/makegroupconfirm.tsx","./src/pages/makegroupstoreinput.tsx","./src/pages/mypage.tsx","./src/pages/onboarding.tsx","./src/pages/register.tsx","./src/pages/restaurantlist.tsx","./src/pages/settingfinished.tsx","./src/pages/splash.tsx","./src/pages/register/registeraccount.tsx","./src/pages/terms/privacypolicy.tsx","./src/pages/terms/serviceterms.tsx","./src/pages/terms/thirdpartypolicy.tsx"],"version":"5.6.3"}
+{"root":["./src/app.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/components/common/agreementbutton.tsx","./src/components/common/backheader.tsx","./src/components/common/bottomnavigation.tsx","./src/components/common/button.tsx","./src/components/common/checkbox.tsx","./src/components/common/flex.tsx","./src/components/common/header.tsx","./src/components/common/searchbar.tsx","./src/components/common/searchdropdown.tsx","./src/components/home/cafe.tsx","./src/components/home/homecard.tsx","./src/components/home/restaurant.tsx","./src/components/home/storeranking.tsx","./src/components/home/taglist.tsx","./src/components/home/titlesection.tsx","./src/components/home/visitedrestaurant.tsx","./src/components/initialsettings/initialsettingscardcontainer.tsx","./src/components/initialsettings/initialsettingsheader.tsx","./src/components/initialsettings/setcafebrandcard.tsx","./src/components/initialsettings/setdepartmentcard.tsx","./src/components/initialsettings/setdrinkfeaturecard.tsx","./src/components/initialsettings/setgendercard.tsx","./src/components/initialsettings/setnamecard.tsx","./src/components/initialsettings/setrestaurantfeaturecard.tsx","./src/components/initialsettings/setrestauranttypecard.tsx","./src/components/makegroup/completestep.tsx","./src/components/makegroup/confirmstep.tsx","./src/components/makegroup/notificationstep.tsx","./src/components/makegroup/participantsstep.tsx","./src/components/makegroup/schedulestep.tsx","./src/components/makegroup/titlestep.tsx","./src/components/mypage/contentbox.tsx","./src/components/restaurantlist/makemuckpotbutton.tsx","./src/components/restaurantlist/muckpotcreatemodal.tsx","./src/components/restaurantlist/muckpotinfobox.tsx","./src/components/restaurantlist/muckpotjoinmodal.tsx","./src/components/restaurantlist/tagfilter.tsx","./src/constants/brands.ts","./src/constants/dummydata.ts","./src/firebase/firebase.ts","./src/pages/cafelist.tsx","./src/pages/findemailandpassword.tsx","./src/pages/group.tsx","./src/pages/home.tsx","./src/pages/initialsettings.tsx","./src/pages/login.tsx","./src/pages/makegroup.tsx","./src/pages/makegroupconfirm.tsx","./src/pages/makegroupstoreinput.tsx","./src/pages/mypage.tsx","./src/pages/onboarding.tsx","./src/pages/register.tsx","./src/pages/restaurantlist.tsx","./src/pages/settingfinished.tsx","./src/pages/splash.tsx","./src/pages/register/registeraccount.tsx","./src/pages/terms/privacypolicy.tsx","./src/pages/terms/serviceterms.tsx","./src/pages/terms/thirdpartypolicy.tsx"],"version":"5.6.3"}
